# cold-reach/api-core/src/openapi_server/models/client.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.client_status import ClientStatus


class Client(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Client - a model defined in OpenAPI

        name: The name of this Client.
        email: The email of this Client.
        company: The company of this Client [Optional].
        status: The status of this Client [Optional].
        last_contact: The last_contact of this Client [Optional].
        id: The id of this Client [Optional].
    """

    name: str = Field(alias="name")
    email: EmailStr = Field(alias="email")
    company: Optional[str] = Field(alias="company", default=None)
    status: Optional[ClientStatus] = Field(alias="status", default=None)
    last_contact: Optional[datetime] = Field(alias="lastContact", default=None)
    id: Optional[str] = Field(alias="id", default=None)

Client.update_forward_refs()