# cold-reach/api-core/src/openapi_server/models/client_status.py
# coding: utf-8

from enum import Enum

class ClientStatus(str, Enum):
    NOT_CONTACTED = "not_contacted"
    EMAILED = "emailed"
    REPLIED = "replied"
    FOLLOW_UP = "follow_up"
    CLOSED = "closed"