# cold-reach/api-core/src/openapi_server/models/email_attachment.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class EmailAttachment(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    EmailAttachment - a model defined in OpenAPI

        id: The id of this EmailAttachment [Optional].
        name: The name of this EmailAttachment [Optional].
        size: The size of this EmailAttachment [Optional].
    """

    id: Optional[str] = Field(alias="id", default=None)
    name: Optional[str] = Field(alias="name", default=None)
    size: Optional[str] = Field(alias="size", default=None)

EmailAttachment.update_forward_refs()