# cold-reach/api-core/src/openapi_server/models/email_request.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.email_attachment import EmailAttachment


class EmailRequest(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    EmailRequest - a model defined in OpenAPI

        recipient_ids: The recipient_ids of this EmailRequest.
        subject: The subject of this EmailRequest.
        body: The body of this EmailRequest.
        template_id: The template_id of this EmailRequest [Optional].
        attachments: The attachments of this EmailRequest [Optional].
    """

    recipient_ids: List[str] = Field(alias="recipient_ids")
    subject: str = Field(alias="subject")
    body: str = Field(alias="body")
    template_id: Optional[str] = Field(alias="template_id", default=None)
    attachment_ids: Optional[List[str]] = Field(alias="attachment_ids", default=None)

EmailRequest.update_forward_refs()