# cold-reach/api-core/src/openapi_server/models/task.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401

class Task(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Task - a model defined in OpenAPI

        id: The id of this Task [Optional].
        type: The type of this Task [Optional].
        client_name: The client_name of this Task [Optional].
        client_email: The client_email of this Task [Optional].
        company: The company of this Task [Optional].
        description: The description of this Task [Optional].
        due_date: The due_date of this Task [Optional].
        priority: The priority of this Task [Optional].
        completed: The completed of this Task [Optional].
    """

    id: Optional[str] = Field(alias="id", default=None)
    type: Optional[str] = Field(alias="type", default=None)
    client_name: Optional[str] = Field(alias="clientName", default=None)
    client_email: Optional[str] = Field(alias="clientEmail", default=None)
    company: Optional[str] = Field(alias="company", default=None)
    description: Optional[str] = Field(alias="description", default=None)
    due_date: Optional[date] = Field(alias="dueDate", default=None)
    priority: Optional[str] = Field(alias="priority", default=None)
    completed: bool = Field(alias="completed", default=False)

    @validator("type")
    def type_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value
        if value not in ('follow_up', 'respond', 'review'):
            raise ValueError("must be one of enum values ('follow_up', 'respond', 'review')")
        return value

    @validator("priority")
    def priority_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value
        if value not in ('high', 'medium', 'low'):
            raise ValueError("must be one of enum values ('high', 'medium', 'low')")
        return value

Task.update_forward_refs()